requirements.txt
azure-identity
azure-keyvault